\name{initialprofile}
\alias{initialprofile}
\title{initialprofile}
\description{
initialprofile 
} 

\usage{
initialprofile(concprofile,limits,slicenumber,dx,cakethickness)}
\arguments{
  \item{concprofile }{
  }
  \item{limits }{
  }
  \item{slicenumber }{
  }
  \item{dx }{
  }
  \item{cakethickness }{
  }
 
}
\value{
  An initial profile
}
\author{
  Pieter Provoost <p.provoost@nioo.knaw.nl>
}
