\name{Luminophore}
\docType{data}
\alias{Luminophore}
\title{Luminophore profile}
\description{
  A profile of Luminophores....
}
\usage{Luminophore}
\format{
  A dataframe with the following columns: .......
}
\examples{

plot(lum~start,data=Luminophore,col=c("red","blue")[Luminophore$profile])

}
\references{??????
}
\keyword{datasets}
