% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regsubsets.fitted.values.R
\name{regsubsets.fitted.values}
\alias{regsubsets.fitted.values}
\title{Fitted Values from Model with \code{nterm} Terms}
\usage{
regsubsets.fitted.values(x, y1, nterm)
}
\arguments{
\item{x}{a numerical vector, viewed as predictors in training sample.}

\item{y1}{a numerical vector, viewed as responses in training sample.}

\item{nterm}{the number of terms to be included in the fitting model.}
}
\value{
\code{regsubsets.fitted.values} produces a vector of fitted values, obtained from a model with \code{nterm} terms.
}
\description{
Fitted values from model with \code{nterm} terms, based on truncated basis functions.
}
\details{
Based on the n × n design matrix for the set of truncated basis functions, \code{regsubsets.fitted.values} computes the fitted values for the model with \code{nterm} terms.
}
\examples{
library(dplyr)
x <- rnorm(50)
y <- sin(x) + rnorm(50,1,2)
nterm <- 7
regsubsets.fitted.values(x,y,nterm)
}
